feat: RatingBar 구현
-drag, tap 제스처 대응

issue: #11

+++ app/src/main/java/com/winteryy/readit/data/local/bookstorage/BookEntity.kt
@@ -18,7 +18,7 @@ data class BookEntity(
     val pubDate: Date,
     val savedDate: Date,
     val bookSaveStatus: BookSaveStatus,
-    val rating: Double
+    val rating: Float
 )
 
 fun BookEntity.toBook() = Book(
